Moved reflection test code into the slang codebase and added the compiler option -reflection-json to store the reflection data in a separate file.
