+++ source/slang/slang-reflection-json.h
@@ -1,0 +1,14 @@
+#ifndef SLANG_REFLECTION_JSON_H
+#define SLANG_REFLECTION_JSON_H
+
+#include "slang.h"
+#include "../core/slang-pretty-writer.h"
+
+namespace Slang
+{
+
+void emitReflectionJSON(SlangCompileRequest* request, SlangReflection* reflection, PrettyWriter& writer);
+
+}
+
+#endif
